Add /store api tests, Add /remote-folder tests

+++ kskp/model.py
@@ -897,6 +897,9 @@ def rename_database_by_id(database_uuid, new_label):
 def delete_database_by_id(database_uuid):
     pass
 
+
+import os
+import pathlib
 import pprint
 from .models.library import Library
 from .models.folder import Folder
@@ -939,6 +942,9 @@ def get_folder2(uuid):
 def set_folder2(f):
     if isinstance(f, Folder):
         library = Library.create_folder_type(f.uuid, f.parent_uuid, f.label, f.creator, f.creator)
+        # フォルダに紐付くディレクトリ(dir_path列で指定されるディレクトリ)がなければ作成する
+        __make_dir(library.dir_path)
+        # libraryレコードをDBに格納する
         library.save()
         f.created_at = library.created_at
     elif isinstance(f, RemoteFolder):
@@ -953,6 +959,9 @@ def set_folder2(f):
                                                   , f.directory
                                                   , f.creator
                                                   , f.creator)
+        # フォルダに紐付くディレクトリ(dir_path列で指定されるディレクトリ)がなければ作成する
+        __make_dir(library.dir_path)
+        # libraryレコードをDBに格納する
         library.save()
         # ここでリモートディレクトリをマウントする
         f.mount(library.dir_path)
@@ -985,11 +994,52 @@ def upd_folder2(f):
 def del_folder2(uuid):
     library = Library.find_by_uuid(uuid)
 
+    if library is None:
+        return
+
     if library.type == 'folder':
         library.delete()
+        __remove_dir(library.dir_path)
     elif library.type == 'remote-folder':
         # ここでリモートディレクトリのマウントを解除する
         remote_folder = RemoteFolder.create_by_library(library)
         remote_folder.unmount(library.dir_path)
         # マウントポイントのディレクトリを削除する
         library.delete()
+        __remove_dir(library.dir_path)
+
+def set_frame2(frame):
+
+    pprint.pprint('>>> 1')
+
+    library = Library.create_frame_type(frame.uuid, frame.parent_uuid, frame.label, frame.creator, frame.creator)
+    # Frameファイルを作成する
+    frame.save(library.dir_path)
+    pprint.pprint('>>> 2')
+    # libraryレコードをDBに格納する
+    library.save()
+    pprint.pprint('>>> 3')
+    frame.created_at = library.created_at
+
+def __make_dir(dir_path):
+    try:
+        dir_path = pathlib.Path(dir_path)
+        # フォルダに紐付くディレクトリ(dir_path列で指定されるディレクトリ)がなければ作成する
+        if dir_path.is_file() or dir_path.is_symlink():
+            raise Exception('Can not make directory, because same name file(%s) exists.' % dir_path)
+        elif not dir_path.is_dir():
+            os.makedirs(str(dir_path), exist_ok=True)
+    except PermissionError as e:
+        # ファイルに対する権限がない場合
+        raise e
+
+def __remove_dir(dir_path):
+    # 全てのフォルダから紐づかないディレクトリは物理削除する
+    dir_path = dir_path.rstrip(os.pathsep)
+    while dir_path != '' and dir_path != '/' and dir_path != 'kskp/data':
+        if Library.dir_path_exists(dir_path):
+            break
+        else:
+            if os.path.isdir(dir_path):
+                os.rmdir(dir_path)
+            dir_path = os.path.dirname(dir_path)
